Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain_mongodb
passlib
python-jose
langchain_groq
langchain_huggingface
longtrainer
pydantic[email]

pdf2image
Pillow
google-genai
python-multipart
huggingface_hub[hf_xet]
groq
